+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### Feat
 
+- offline US federal holidays; move scraper to Web class
 - dehydrate stpstone calendars into wwdates
 - first commit